refactor: migrate TimeUtilities, JobStatus, and StateMachine to DIRACCommon
Migrate three core modules from DIRAC to DIRACCommon to enable shared usage:

- Move TimeUtilities.py with gLogger references to DIRACCommon/Core/Utilities/
- Move StateMachine.py with State and StateMachine classes to DIRACCommon/Core/Utilities/
- Move JobStatus.py with constants and JobsStateMachine to DIRACCommon/WorkloadManagementSystem/Client/

DIRAC modules now re-export from DIRACCommon for backward compatibility.
Update JobStatusUtility.py imports to use migrated DIRACCommon modules.
