+++ dirac-common/src/DIRACCommon/WorkloadManagementSystem/Client/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon WorkloadManagementSystem client utilities"""
